requirements.txt
pandas
numpy
matplotlib
scikit-learn
seaborn
xgboost
kagglehub
scipy
streamlit
python-dotenv